+++ src/test/java/org/terasology/polyworld/sampling/PoissonDiscTestView.java
@@ -30,7 +30,7 @@ import javax.swing.JPanel;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.math.Rect2i;
+import org.terasology.math.geom.Rect2f;
 import org.terasology.math.geom.Vector2f;
 import org.terasology.math.geom.Vector2i;
 
@@ -49,7 +49,7 @@ public final class PoissonDiscTestView {
 //      int rows = DoubleMath.roundToInt(area.height() / cellSize, RoundingMode.HALF_UP);
 //      int cols = DoubleMath.roundToInt(area.width() / cellSize, RoundingMode.HALF_UP);
 
-        Rect2i area = Rect2i.createFromMinAndSize(30, 10, 512, 256);
+        Rect2f area = Rect2f.createFromMinAndSize(30, 10, 512, 256);
         int numSites = DoubleMath.roundToInt(area.area() * graphDensity / 1000, RoundingMode.HALF_UP);
 
         logger.info("START GRID");
@@ -86,7 +86,7 @@ public final class PoissonDiscTestView {
                 }
             }
 
-            private void drawDebug(Rect2i bounds, Graphics2D g) {
+            private void drawDebug(Rect2f bounds, Graphics2D g) {
                 g.setColor(Color.LIGHT_GRAY);
                 Vector2i dims = sampling.getGridDimensions(bounds, numSites);
                 int cols = dims.getX();
@@ -97,10 +97,10 @@ public final class PoissonDiscTestView {
 
                 g.translate(bounds.minX(), bounds.minY());
                 for (int i = 0; i <= cols; i++) {
-                    g.drawLine((int) (i * cellWidth), 0, (int) (i * cellWidth), bounds.height());
+                    g.drawLine((int) (i * cellWidth), 0, (int) (i * cellWidth), (int) bounds.height());
                 }
                 for (int i = 0; i <= rows; i++) {
-                    g.drawLine(0, (int) (i * cellHeight), bounds.width(), (int) (i * cellHeight));
+                    g.drawLine(0, (int) (i * cellHeight), (int) bounds.width(), (int) (i * cellHeight));
                 }
                 g.translate(-bounds.minX(), -bounds.minY());
             }
